// src/api/calculator/models/Pressure.ts
// tslint:disable
/**
 * h2AuxCalculatorApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2018-05-24T17:52:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface Pressure
 */
export interface Pressure {
    /**
     * @type {number}
     * @memberof Pressure
     */
    value?: number;
    /**
     * @type {string}
     * @memberof Pressure
     */
    unit?: PressureUnitEnum;
}

/**
 * @export
 * @enum {string}
 */
export enum PressureUnitEnum {
    Bar = 'BAR',
    Psi = 'PSI'
}